requirements.txt
affine==2.4.0
aspose-cells==23.5.0
astroid==2.3.3
async-generator==1.10
attrs==19.3.0
autopep8==1.6.0
backcall==0.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==3.1.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
clarifai==2.6.2
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cmake==3.20.2
colorama==0.4.3
configparser==3.8.1
cryptography==36.0.1
customtkinter==5.1.3
cycler==0.10.0
darkdetect==0.8.0
decorator==4.4.1
defusedxml==0.6.0
dnspython==2.1.0
docopt==0.6.2
EasyProcess==1.1
emoji==1.6.3
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
face-recognition-models==0.3.0
Flask==2.0.1
future==0.18.2
GDAL==3.4.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
h11==0.13.0
idna==3.3
imageio==2.26.1
importlib-metadata==1.5.0
imutils==0.5.4
iniconfig==2.0.0
instapy==0.2.3
instapy-chromedriver==2.46
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==2.0.1
jedi==0.16.0
Jinja2==3.0.1
joblib==1.2.0
JPype1==1.4.1
Js2Py==0.74
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
MeaningCloud-python==2.0.0
mistune==0.8.4
mysql-connector==2.2.9
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.6.3
notebook==6.0.3
numpy==1.20.3
opencv-python==4.5.2.52
openpyxl==3.1.2
outcome==1.1.0
packaging==23.0
paho-mqtt==1.5.1
pandas==1.3.5
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==8.2.0
pip==23.1.2
pipwin==0.5.2
pluggy==1.0.0
plyer==2.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.19.4
pycodestyle==2.8.0
pycparser==2.21
pygame==1.9.5
Pygments==2.5.2
pyjsparser==2.7.1
pylint==2.4.4
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyPrind==2.11.3
pyrsistent==0.15.7
pySmartDL==1.3.4
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.1
python-etcd==0.4.5
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==3.0
PyWavelets==1.3.0
pywin32==227
pywinpty==0.5.7
PyYAML==6.0
pyzmq==19.0.0
qtconsole==4.6.0
rasterio==1.2.10
regex==2022.3.2
requests==2.27.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.0
selenium==4.1.2
Send2Trash==1.5.0
setuptools==40.8.0
shapely==2.0.1
simplegeneric==0.8.1
six==1.14.0
sniffio==1.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==3.1.0
tifffile==2021.11.2
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.0.3
tqdm==4.61.1
traitlets==4.3.3
trio==0.20.0
trio-websocket==0.9.2
typed-ast==1.4.1
typing_extensions==4.1.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.38.4
widgetsnbextension==3.5.1
wrapt==1.11.2
wsproto==1.1.0
zipp==3.0.0
